Updating for subscription API tests
